Rename LeadershipNew to Leadership.tsx and leadership to leadershipProfiles.ts
